CMake and file structure
